requirements.txt
Django==5.0.6
django-admin-sortable2==2.2.1
django-axes==6.4.0
django-bitfield==2.2.0
django-bootstrap-modal-forms==3.0.4
django-bootstrap5==24.2
django-chartjs==2.3.0
django-filter==24.2
django-formtools==2.5.1
django-mail-templated==2.6.5
# django-two-factor-auth depends on
# django-phonenumber-field < 7.0.0
django-phonenumber-field==6.4.0
django-two-factor-auth==1.16.0
phonenumbers==8.13.37
psycopg2-binary==2.9.9
